[MOD] Modified delete function to from null trown IllegalArgumentException [TEST] all test passed

+++ activity-tracker-persistence/src/main/java/cz/muni/fi/pa165/tracker/dao/ActivityReportDaoImpl.java
@@ -50,6 +50,9 @@ public class ActivityReportDaoImpl implements ActivityReportDao {
 
     @Override
     public void delete(ActivityReport activityReport) {
+        if (activityReport == null) {
+            throw new IllegalArgumentException("Activity report like null is not valid");
+        }
         em.remove(findActivityReportByID(activityReport.getId()));
     }
