+++ activity-tracker-persistence/src/main/java/cz/muni/fi/pa165/tracker/dao/UserDaoImpl.java
@@ -57,6 +57,9 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     public void remove(User user) {
+        if (user == null) {
+            throw new IllegalArgumentException("Usetr like null is not valid");
+        }
         em.remove(findById(user.getId()));
     }
 }
